Added support for non-RGB Color Modes
Personal update I did to get the script working properly with images that use Indexed color mode.

In isometric-tile.lua I shuffled a lot of stuff around and made an attempt at tidying up the code a bit. It's my first time working with lua scripts and the aseprite API so it helped me familiarize and learn my way around - and identify what was going wrong (before I found that the bug was related to Color Modes).

Once i was happy with the "tile" script I was a lot lazier about updating the remaining scripts I just sort of threw the extra ColorMode arguments in there to get them working ~